Move Helm dependency manager to `helm` package
