+++ internal/helm/testdata/charts/helmchart/templates/serviceaccount.yaml
@@ -1,0 +1,8 @@
+{{- if .Values.serviceAccount.create -}}
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ include "helmchart.serviceAccountName" . }}
+  labels:
+{{ include "helmchart.labels" . | nindent 4 }}
+{{- end -}}
