+++ internal/helm/testdata/charts/helmchart/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "helmchart.fullname" . }}-test-connection"
+  labels:
+{{ include "helmchart.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test-success
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args:  ['{{ include "helmchart.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
