+++ internal/helm/testdata/charts/helmchartwithdeps/templates/service.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: {{ include "helmchartwithdeps.fullname" . }}
+  labels:
+    {{- include "helmchartwithdeps.labels" . | nindent 4 }}
+spec:
+  type: {{ .Values.service.type }}
+  ports:
+    - port: {{ .Values.service.port }}
+      targetPort: http
+      protocol: TCP
+      name: http
+  selector:
+    {{- include "helmchartwithdeps.selectorLabels" . | nindent 4 }}
