+++ internal/helm/testdata/charts/helmchartwithdeps/templates/serviceaccount.yaml
@@ -1,0 +1,8 @@
+{{- if .Values.serviceAccount.create -}}
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: {{ include "helmchartwithdeps.serviceAccountName" . }}
+  labels:
+{{ include "helmchartwithdeps.labels" . | nindent 4 }}
+{{- end -}}
