+++ internal/helm/testdata/charts/helmchartwithdeps/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "helmchartwithdeps.fullname" . }}-test-connection"
+  labels:
+{{ include "helmchartwithdeps.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test-success
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args:  ['{{ include "helmchartwithdeps.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
